requirements.txt
langchain==0.3.27
python-dotenv==1.0.0
streamlit==1.37.0
pandas==2.1.4

langchain_community==0.3.27
langchain-core==0.3.74
langchain-ollama~=0.3.6

langgraph==0.6.4

matplotlib==3.10.5
graphviz==0.20.1
